requirements.txt
py-cord==2.6.1
firebase-admin==6.5.0
python-dotenv==1.0.1
pytz==2024.1
